+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 ### Fixed
 - Fixed a problem where new SSIDs weren't shown when the filter function was used right after the list of SSIDs was updated.
 
+### Removed
+- The feature to add own SSIDs was removed. There should be a folder named "freifunkautoconnect" containing the file "user_ssids.csv" on your SD card if version <= 0.5 was installed on your device. You can remove this folder, it is no longer needed.
+
 ## [0.5] 2015-08-31
 ### Added
 - The app will check on start up, if an updated list of SSIDs is available. If the check didn't fail, the next check will not be started until 24 hours have passed.